feat: REQ-1912 Se agrego la funcionalidad para el stock por sucursales, se refactorizo el flujo de ventas y compras. Se agregaron los registro de usuario y se cambiaron las clases por enums

+++ src/main/java/com/svc/ventas/message/request/EmpleadoCreateRequest.java
@@ -16,6 +16,6 @@ public class EmpleadoCreateRequest {
   private Integer idPersona;
 
   @NotNull
-  private Integer idSucursal;
+  private Long idSucursal;
 
 }
