+++ src/main/java/com/svc/ventas/models/dao/SucursalRepo.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.svc.ventas.models.entity.Sucursal;
 
-public interface SucursalRepo extends JpaRepository<Sucursal, Integer>{
+public interface SucursalRepo extends JpaRepository<Sucursal, Long>{
 
 	@Query("SELECT s FROM Sucursal  s WHERE s.indEstado= true")
 	List<Sucursal> findSucursales();
