+++ src/main/java/com/svc/ventas/models/specifications/ProductSpecifications.java
@@ -1,9 +1,6 @@
 package com.svc.ventas.models.specifications;
 
-import com.svc.ventas.models.entity.Categoria;
 import com.svc.ventas.models.entity.Producto;
-import jakarta.persistence.criteria.Join;
-import jakarta.persistence.criteria.JoinType;
 import org.springframework.data.jpa.domain.Specification;
 
 public class ProductSpecifications {
@@ -15,16 +12,12 @@ public class ProductSpecifications {
                     criteriaBuilder.like(criteriaBuilder.lower(root.get("codigo")), "%" + name.toLowerCase() + "%"));
   }
 
-  public static Specification<Producto> hasCategory(Long categoryId) {
+  public static Specification<Producto> hasCategory(Integer categoryId) {
     return (root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get("categoria").get("idCategoria"), categoryId);
   }
 
   public static Specification<Producto> hasStatus(Boolean status) {
     return (root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get("indEstado"), status);
   }
-
-
-
-
 
 }
